+++ src/Microsoft.Health.Fhir.SqlServer/Features/Schema/Sql/Tables/TokenText.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE dbo.TokenText
 (
-    ResourceTypeId              smallint            NOT NULL,
+    ResourceTypeId              tinyint            NOT NULL,
     ResourceSurrogateId         bigint              NOT NULL,
     SearchParamId               smallint            NOT NULL,
     Text                        nvarchar(400)       COLLATE Latin1_General_CI_AI NOT NULL,
