+++ src/Microsoft.Health.Fhir.SqlServer/Features/Storage/ResourceWrapperExtention.cs
@@ -10,7 +10,7 @@ namespace Microsoft.Health.Fhir.SqlServer.Features.Storage
 {
     public static class ResourceWrapperExtention
     {
-        public static ResourceDateKey ToResourceDateKey(this ResourceWrapper wrapper, Func<string, short> getResourceTypeId, bool ignoreVersion = false)
+        public static ResourceDateKey ToResourceDateKey(this ResourceWrapper wrapper, Func<string, byte> getResourceTypeId, bool ignoreVersion = false)
         {
             return new ResourceDateKey(getResourceTypeId(wrapper.ResourceTypeName), wrapper.ResourceId, wrapper.LastModified.ToSurrogateId(), ignoreVersion ? null : wrapper.Version);
         }
