+++ src/Microsoft.Health.Fhir.SqlServer/Features/Storage/TvpRowGeneration/Merge/TokenNumberNumberCompositeSearchParamListRowGenerator.cs
@@ -22,7 +22,7 @@ namespace Microsoft.Health.Fhir.SqlServer.Features.Storage.TvpRowGeneration
         }
 
         internal override bool TryGenerateRow(
-            short resourceTypeId,
+            byte resourceTypeId,
             long resourceSurrogateId,
             short searchParamId,
             (TokenSearchValue component1, NumberSearchValue component2, NumberSearchValue component3) searchValue,
